Triad Finale
The end of the fight is nearly complete

Trails namespace is now VFX, intended all projectiles that are primarily for that purpose

Also nerfed Armored Vikings lol

+++ Projectiles/Enemy/EnemyGenericLaser.cs
@@ -300,7 +300,7 @@ namespace tsorcRevamp.Projectiles.Enemy
             {
                 if (FiringTimeLeft < fadeOutFrames)
                 {
-                    fadePercent = FiringTimeLeft / fadeOutFrames;
+                    fadePercent = (float)FiringTimeLeft / (float)fadeOutFrames;
                 }
                 else
                 {
